Refactoring of tagOccurences(), respect ACLs and hidden pages
tagOccurences does now only return pages that are visible and readable
for the current user. If there is any use for getting all pages (i.e.
also invisible/unreadable ones) this choice might be added as parameter.

The code has been rewritten in order to make it easier to understand and
the matching of recursive namespaces has been fixed.