Disable caching for the count syntax, load occurrences only in xhtml mode
As the count syntax does respect ACLs now, its output needs to be
adjusted for the current user. As no caching logic for new/changed pages
or changes in the permissions of the user that affect the output of the
count syntax has been implemented yet, the cache has been simply
deactivated.