New version with metadata index
This version of the tag plugin now uses the metadata index of DokuWiki
as tag storage. The old tag database is no longer needed. Plugins that
use the tag plugin as helper like the cloud plugin need to be updated,
too. The metadata index can be rebuilt using the searchindex plugin or
bin/indexer.php, the old admin interface no longer exists.

Tags that are normalized to the same page id are now the same, you can
control this normalization using the deaccent configuration option.