Remove deletion of persistent tag data
This makes the tag plugin more compatible with blogtng in the sense that
you can specify both tags in the page itself as well as in blogtng and
both will be combined.

Persistent tag data was previously deleted because very old versions of
the tag plugin created persistent metadata which couldn't be removed
anymore.