New feature: a form for searching pages by tags
This introduces a new syntax {{searchtags}} that displays a search form
with the possibility to select a namespace and to select tags from a
table of all tags that should either be in- or excluded with the option
to choose if the tags should be associated with AND or OR. The results
are displayed as in the topic syntax, in the form
{{searchtags&flag[&flag..]}} it is possible to specify the same flags as
for the topic syntax.

The implementation of this feature has been paid by the Bright Group
(bright-group.ru).