Revert the tag query change that was introduced in version 2012-08-24
The semantic of a tag query is relatively obscure but also clearly
documented. Each part of the query relates to the result of the previous
parts, tags without prefix add pages with that tag to the result, a tag
prefixed with "+" removes pages without that tag, a tag prefixed with
"-" removes pages with that tag.

After the 2012-08-24 version this behavior was a bit different, here the
query was executed as if all prefixed tags were at the end.

With this change the original behavior is restored.