Assume an implicit query for all pages if no tag query is given, fixes #35
This lets the tag plugin assume an implicit query for all pages if the
query doesn't start with an unprefixed tag. This means that now "+tag1
+tag2" returns the same results as "tag1 +tag2" and "-tag1" returns all
pages that are not tagged with "tag1". This also adds tests for this
behavior. The tagRefine-function now uses the same code as the
getTopic-function except that the result is limited to the specified
pages and that no filters apart from the tags are applied.