fixes issue 145: Add sort key flag to topic
Issue 145 addresses a problem a came across recently:
topics are sorted "sortkey" and "sortorder" according to settings in plugin config. 
Changing this "on demand" might be favourable sometimes. In my case, default sorting in config is set to "title", but in a namespace called "news" i would like to sort by date with newest items first.
This patch lets you add the flags "sortkey" and "sortorder", preserving all other functionality: 
{{topic>.?news&table&header&comments&sortkey=mdate&sortorder=descending}}